+++ R/RapResults.R
@@ -30,7 +30,7 @@ NULL
 #' \describe{
 #' \item{Run_Number}{The index of each solution in the object.}
 #' \item{Status}{The status of the solution. The values in this column
-#' correspond to outputs from the Gurobi software package (<http://www.gurobi.com/documentation/6.5/refman/optimization_status_codes.html>).}
+#' correspond to outputs from the Gurobi software package (<https://www.gurobi.com/documentation/6.5/refman/optimization_status_codes.html>).}
 #' \item{Score}{The objective function for the solution.}
 #' \item{Cost}{Total cost associated with a solution.}
 #' \item{Planning_Units}{Number of planning units selected in a solution.}
@@ -151,7 +151,7 @@ methods::setClass("RapResults",
 #' \describe{
 #' \item{Run_Number}{The index of each solution in the object.}
 #' \item{Status}{The status of the solution. The values in this column
-#' correspond to outputs from the Gurobi software package (<http://www.gurobi.com/documentation/6.5/refman/optimization_status_codes.html>).}
+#' correspond to outputs from the Gurobi software package (<https://www.gurobi.com/documentation/6.5/refman/optimization_status_codes.html>).}
 #' \item{Score}{The objective function for the solution.}
 #' \item{Cost}{Total cost associated with a solution.}
 #' \item{Planning_Units}{Number of planning units selected in a solution.}